Move permissions.py from controllers to main
